@@ -1,6 +1,6 @@
 {
     "name": "tochka-developers/jsonrpc",
-    "description": "JsonRpc extension for Laravel",
+    "description": "JsonRpc server package for Laravel",
     "license": "MIT",
     "authors": [
         {
@@ -13,37 +13,47 @@
         "laravel"
     ],
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.0",
         "ext-json": "*",
-        "bensampo/laravel-enum": "^1.0|^2.0|^3.0|^4.0|^5.0,<5.2.0",
-        "illuminate/support": "^6.0|^7.0|^8.0|^9.0",
-        "illuminate/http": "^6.0|^7.0|^8.0|^9.0",
-        "illuminate/container": "^6.0|^7.0|^8.0|^9.0",
-        "illuminate/pipeline": "^6.0|^7.0|^8.0|^9.0",
-        "illuminate/console": "^6.0|^7.0|^8.0|^9.0",
+        "doctrine/annotations": "^1.11",
+        "illuminate/contracts": "^9.0",
+        "illuminate/support": "^9.0",
+        "illuminate/http": "^9.0",
+        "illuminate/container": "^9.0",
+        "illuminate/pipeline": "^9.0",
+        "illuminate/console": "^9.0",
+        "illuminate/validation": "^9.0",
         "phpdocumentor/reflection-docblock": "^3.0|^4.0|^5.0",
+        "phpdocumentor/type-resolver": "^1.5",
+        "psr/http-message": "^1.0",
+        "psr/simple-cache": "^1.0|^2.0|^3.0",
         "spiral/attributes": "^2.8",
-        "tochka-developers/jsonrpc-annotations": "^1.1",
-        "psr/simple-cache": "^1.0|^2.0"
+        "symfony/http-foundation": "^6.0",
+        "symfony/http-kernel": "^6.0",
+        "tochka-developers/array-file-cache": "^1.0|^2.0|^3.0",
+        "tochka-developers/jsonrpc-annotations": "^1.3",
+        "tochka-developers/jsonrpc-standard": "^1.0"
     },
     "require-dev": {
+        "bensampo/laravel-enum": "^5.0",
+        "laravel/pint": "^1.4",
+        "mockery/mockery": "^1.0",
+        "orchestra/testbench": "^7.1",
+        "phpunit/phpunit": "^9.6",
         "roave/security-advisories": "dev-latest",
-        "phpunit/phpunit": "^9.5",
-        "mockery/mockery": "^1.0"
+        "timacdonald/log-fake": "^2.0",
+        "vimeo/psalm": "^5.6"
     },
     "minimum-stability": "dev",
     "prefer-stable": true,
     "autoload": {
         "psr-4": {
             "Tochka\\JsonRpc\\": "src/"
-        },
-        "files": [
-            "src/helpers.php"
-        ]
+        }
     },
     "autoload-dev": {
         "psr-4": {
-            "Tochka\\JsonRpc\\Tests\\": "tests/"
+            "Tochka\\JsonRpc\\Tests\\": "tests"
         }
     },
     "extra": {
@@ -52,5 +62,10 @@
                 "Tochka\\JsonRpc\\JsonRpcServiceProvider"
             ]
         }
+    },
+    "scripts": {
+        "lint": "php8.1 vendor/bin/pint --test",
+        "lint-fix": "php8.1 vendor/bin/pint",
+        "analyze": "php8.1 vendor/bin/psalm --php-version=8.1 --no-cache"
     }
 }
