@@ -2,7 +2,9 @@
 
 namespace Tochka\JsonRpc\Contracts;
 
+use Tochka\JsonRpc\Route\Parameters\Parameter;
+
 interface CustomCasterInterface
 {
-    public function cast(string $expectedType, $value, string $fieldName);
+    public function cast(Parameter $parameter, mixed $value, string $fieldName): ?object;
 }
