@@ -12,10 +12,15 @@ interface ParamsResolverInterface
      * @return array<string, Parameter>
      */
     public function resolveParameters(\ReflectionMethod $reflectionMethod): array;
-    
+
     public function resolveResult(\ReflectionMethod $reflectionMethod): Parameter;
-    
+
+    /** @return array<string, ParameterObject> */
     public function getClasses(): array;
-    
+
+    /**
+     * @param class-string $className
+     * @return ParameterObject|null
+     */
     public function getParameterObject(string $className): ?ParameterObject;
 }
